requirements.txt
langchain-core
langchain-groq==0.3.2
python-dotenv==1.1.0
pypdf==5.4.0
python-docx==1.1.2
